# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title
  title: Api
  version: 1.0.0
  description: Cricket League Auction & Bidding Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Unauthorized
...
  /auth/users:
    get:
      operationId: listUsers
      tags: [auth]
      summary: List all users (admin)
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
    post:
      operationId: createUser
      tags: [auth]
      summary: Create user (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

  # PLAYERS
  /players:
    get:
      operationId: listPlayers
      tags: [players]
      summary: List all players
      parameters:
        - name: category
          in: query
          schema:
            type: string
        - name: country
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of players
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Player"
    post:
      operationId: createPlayer
      tags: [players]
      summary: Create player
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlayerInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
  /players/{id}:
    get:
      operationId: getPlayer
      tags: [players]
      summary: Get player by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Player
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
        "404":
          description: Not found
    patch:
      operationId: updatePlayer
      tags: [players]
      summary: Update player
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlayerUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
    delete:
      operationId: deletePlayer
      tags: [players]
      summary: Delete player
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # TEAMS
  /teams:
    get:
      operationId: listTeams
      tags: [teams]
      summary: List all teams
      responses:
        "200":
          description: List of teams
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Team"
    post:
      operationId: createTeam
      tags: [teams]
      summary: Create team
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
  /teams/{id}:
    get:
      operationId: getTeam
      tags: [teams]
      summary: Get team by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Team
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
    patch:
      operationId: updateTeam
      tags: [teams]
      summary: Update team
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Team"
  /teams/{id}/squad:
    get:
      operationId: getTeamSquad
      tags: [teams]
      summary: Get team squad
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Team squad
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamSquad"

  # AUCTIONS
...
components:
  schemas:
...
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      required: [id, email, name, role, createdAt]
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [admin, commissioner, auctioneer, team_owner, viewer, analyst]
        teamId:
          type: ["integer", "null"]
        createdAt:
          type: string
    UserInput:
      type: object
      required: [email, name, role, password]
      properties:
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [admin, commissioner, auctioneer, team_owner, viewer, analyst]
        password:
          type: string
        teamId:
          type: ["integer", "null"]

    # PLAYERS
    Player:
      type: object
      required: [id, name, country, category, basePrice, battingStyle, bowlingStyle, status, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        country:
          type: string
        city:
          type: ["string", "null"]
        age:
          type: ["integer", "null"]
        category:
          type: string
          enum: [batsman, bowler, all_rounder, wicket_keeper]
        battingStyle:
          type: string
        bowlingStyle:
          type: string
        basePrice:
          type: number
        status:
          type: string
          enum: [available, sold, unsold, in_auction]
        photoUrl:
          type: ["string", "null"]
        matches:
          type: ["integer", "null"]
        runs:
          type: ["integer", "null"]
        wickets:
          type: ["integer", "null"]
        strikeRate:
          type: ["number", "null"]
        economy:
          type: ["number", "null"]
        average:
          type: ["number", "null"]
        fifties:
          type: ["integer", "null"]
        hundreds:
          type: ["integer", "null"]
        rating:
          type: ["number", "null"]
        tags:
          type: array
          items:
            type: string
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
    PlayerInput:
      type: object
      required: [name, country, category, basePrice, battingStyle, bowlingStyle]
      properties:
        name:
          type: string
        country:
          type: string
        city:
          type: string
        age:
          type: integer
        category:
          type: string
          enum: [batsman, bowler, all_rounder, wicket_keeper]
        battingStyle:
          type: string
        bowlingStyle:
          type: string
        basePrice:
          type: number
        photoUrl:
          type: string
        matches:
          type: integer
        runs:
          type: integer
        wickets:
          type: integer
        strikeRate:
          type: number
        economy:
          type: number
        average:
          type: number
        fifties:
          type: integer
        hundreds:
          type: integer
        rating:
          type: number
        tags:
          type: array
          items:
            type: string
        notes:
          type: string
    PlayerUpdate:
      type: object
      properties:
        name:
          type: string
        country:
          type: string
        city:
          type: string
        age:
          type: integer
        category:
          type: string
          enum: [batsman, bowler, all_rounder, wicket_keeper]
        battingStyle:
          type: string
        bowlingStyle:
          type: string
        basePrice:
          type: number
        photoUrl:
          type: string
        matches:
          type: integer
        runs:
          type: integer
        wickets:
          type: integer
        strikeRate:
          type: number
        economy:
          type: number
        average:
          type: number
        fifties:
          type: integer
        hundreds:
          type: integer
        rating:
          type: number
        tags:
          type: array
          items:
            type: string
        notes:
          type: string
        status:
          type: string
          enum: [available, sold, unsold, in_auction]

    # TEAMS
    Team:
      type: object
      required: [id, name, shortName, primaryColor, purse, remainingPurse, maxPlayers, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        shortName:
          type: string
        primaryColor:
          type: string
        secondaryColor:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
        purse:
          type: number
        remainingPurse:
          type: number
        maxPlayers:
          type: integer
        maxOverseas:
          type: integer
        ownerId:
          type: ["integer", "null"]
        ownerName:
          type: ["string", "null"]
        createdAt:
          type: string
    TeamInput:
      type: object
      required: [name, shortName, primaryColor, purse, maxPlayers]
      properties:
        name:
          type: string
        shortName:
          type: string
        primaryColor:
          type: string
        secondaryColor:
          type: string
        logoUrl:
          type: string
        purse:
          type: number
        maxPlayers:
          type: integer
        maxOverseas:
          type: integer
        ownerId:
          type: integer
        ownerName:
          type: string
    TeamUpdate:
      type: object
      properties:
        name:
          type: string
        shortName:
          type: string
        primaryColor:
          type: string
        secondaryColor:
          type: string
        logoUrl:
          type: string
        purse:
          type: number
        remainingPurse:
          type: number
        maxPlayers:
          type: integer
        maxOverseas:
          type: integer
        ownerId:
          type: integer
        ownerName:
          type: string
    TeamSquad:
      type: object
      required: [team, players, stats]
      properties:
        team:
          $ref: "#/components/schemas/Team"
        players:
          type: array
          items:
            $ref: "#/components/schemas/SquadPlayer"
        stats:
          $ref: "#/components/schemas/SquadStats"
    SquadPlayer:
      type: object
      required: [id, name, category, country, soldPrice, battingStyle, bowlingStyle]
      properties:
        id:
          type: integer
        name:
          type: string
        category:
          type: string
        country:
          type: string
        soldPrice:
          type: number
        battingStyle:
          type: string
        bowlingStyle:
          type: string
        photoUrl:
          type: ["string", "null"]
        rating:
          type: ["number", "null"]
    SquadStats:
      type: object
      required: [totalPlayers, batsmen, bowlers, allRounders, wicketKeepers, overseas, spent, remaining, avgPrice]
      properties:
        totalPlayers:
          type: integer
        batsmen:
          type: integer
        bowlers:
          type: integer
        allRounders:
          type: integer
        wicketKeepers:
          type: integer
        overseas:
          type: integer
        spent:
          type: number
        remaining:
          type: number
        avgPrice:
          type: number

    # AUCTIONS
...