# lib/api-zod/src/generated/types/squadStats.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Cricket League Auction & Bidding Platform API
 * OpenAPI spec version: 1.0.0
 */

export interface SquadStats {
  totalPlayers: number;
  batsmen: number;
  bowlers: number;
  allRounders: number;
  wicketKeepers: number;
  overseas: number;
  spent: number;
  remaining: number;
  avgPrice: number;
}