# lib/api-zod/src/generated/types/teamSquad.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Cricket League Auction & Bidding Platform API
 * OpenAPI spec version: 1.0.0
 */
import type { SquadPlayer } from "./squadPlayer";
import type { SquadStats } from "./squadStats";
import type { Team } from "./team";

export interface TeamSquad {
  team: Team;
  players: SquadPlayer[];
  stats: SquadStats;
}