# lib/api-zod/src/generated/types/teamUpdate.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Cricket League Auction & Bidding Platform API
 * OpenAPI spec version: 1.0.0
 */

export interface TeamUpdate {
  name?: string;
  shortName?: string;
  primaryColor?: string;
  secondaryColor?: string;
  logoUrl?: string;
  purse?: number;
  remainingPurse?: number;
  maxPlayers?: number;
  maxOverseas?: number;
  ownerId?: number;
  ownerName?: string;
}