# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Cricket League Auction & Bidding Platform API
 * OpenAPI spec version: 1.0.0
 */
import type { UserRole } from "./userRole";

export interface User {
  id: number;
  email: string;
  name: string;
  role: UserRole;
  /** @nullable */
  teamId?: number | null;
  createdAt: string;
}